+++ plugin-gradle/CHANGES.md
@@ -3,6 +3,7 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `3.27.0`).
 
 ## [Unreleased]
+* Support for ktfmt in KotlinGradleExtension ([#582](https://github.com/diffplug/spotless/pull/582))
 
 ## [4.0.1] - 2020-05-21
 ### Fixed